further split up tests
