+++ test/fields/e2e.spec.ts
@@ -10,12 +10,8 @@ import type { Config } from './payload-types.js'
 
 import {
   ensureAutoLoginAndCompilationIsDone,
-  exactText,
   initPageConsoleErrorCatch,
-  navigateToListCellLink,
-  openDocDrawer,
   saveDocAndAssert,
-  switchTab,
 } from '../helpers.js'
 import { AdminUrlUtil } from '../helpers/adminUrlUtil.js'
 import { initPayloadE2ENoConfig } from '../helpers/initPayloadE2ENoConfig.js'
@@ -23,16 +19,7 @@ import { reInitializeDB } from '../helpers/reInitializeDB.js'
 import { RESTClient } from '../helpers/rest.js'
 import { POLL_TOPASS_TIMEOUT, TEST_TIMEOUT_LONG } from '../playwright.config.js'
 import { jsonDoc } from './collections/JSON/shared.js'
-import { numberDoc } from './collections/Number/shared.js'
-import { textDoc } from './collections/Text/shared.js'
-import {
-  arrayFieldsSlug,
-  blockFieldsSlug,
-  collapsibleFieldsSlug,
-  pointFieldsSlug,
-  tabsFieldsSlug,
-  textFieldsSlug,
-} from './slugs.js'
+import { arrayFieldsSlug, blockFieldsSlug, collapsibleFieldsSlug } from './slugs.js'
 const filename = fileURLToPath(import.meta.url)
 const dirname = path.dirname(filename)
 
@@ -77,242 +64,6 @@ describe('fields', () => {
     await client.login()
 
     await ensureAutoLoginAndCompilationIsDone({ page, serverURL })
-  })
-  describe('text', () => {
-    let url: AdminUrlUtil
-    beforeAll(() => {
-      url = new AdminUrlUtil(serverURL, textFieldsSlug)
-    })
-
-    test('should display field in list view', async () => {
-      await page.goto(url.list)
-      const textCell = page.locator('.row-1 .cell-text')
-      await expect(textCell).toHaveText(textDoc.text)
-    })
-
-    test('should hide field in column selector when admin.disableListColumn', async () => {
-      await page.goto(url.list)
-      await page.locator('.list-controls__toggle-columns').click()
-
-      await expect(page.locator('.column-selector')).toBeVisible()
-
-      // Check if "Disable List Column Text" is not present in the column options
-      await expect(
-        page.locator(`.column-selector .column-selector__column`, {
-          hasText: exactText('Disable List Column Text'),
-        }),
-      ).toBeHidden()
-    })
-
-    test('should show field in filter when admin.disableListColumn is true', async () => {
-      await page.goto(url.list)
-      await page.locator('.list-controls__toggle-where').click()
-      await page.locator('.where-builder__add-first-filter').click()
-
-      const initialField = page.locator('.condition__field')
-      await initialField.click()
-
-      await expect(
-        initialField.locator(`.rs__menu-list:has-text("Disable List Column Text")`),
-      ).toBeVisible()
-    })
-
-    test('should display field in list view column selector if admin.disableListColumn is false and admin.disableListFilter is true', async () => {
-      await page.goto(url.list)
-      await page.locator('.list-controls__toggle-columns').click()
-
-      await expect(page.locator('.column-selector')).toBeVisible()
-
-      // Check if "Disable List Filter Text" is present in the column options
-      await expect(
-        page.locator(`.column-selector .column-selector__column`, {
-          hasText: exactText('Disable List Filter Text'),
-        }),
-      ).toBeVisible()
-    })
-
-    test('should hide field in filter when admin.disableListFilter is true', async () => {
-      await page.goto(url.list)
-      await page.locator('.list-controls__toggle-where').click()
-      await page.locator('.where-builder__add-first-filter').click()
-
-      const initialField = page.locator('.condition__field')
-      await initialField.click()
-
-      await expect(
-        initialField.locator(`.rs__option :has-text("Disable List Filter Text")`),
-      ).toBeHidden()
-    })
-
-    test('should display i18n label in cells when missing field data', async () => {
-      await page.goto(url.list)
-      const textCell = page.locator('.row-1 .cell-i18nText')
-      await expect(textCell).toHaveText('<No Text en>')
-    })
-
-    test('should show i18n label', async () => {
-      await page.goto(url.create)
-
-      await expect(page.locator('label[for="field-i18nText"]')).toHaveText('Text en')
-    })
-
-    test('should show i18n placeholder', async () => {
-      await page.goto(url.create)
-      await expect(page.locator('#field-i18nText')).toHaveAttribute('placeholder', 'en placeholder')
-    })
-
-    test('should show i18n descriptions', async () => {
-      await page.goto(url.create)
-      const description = page.locator('.field-description-i18nText')
-      await expect(description).toHaveText('en description')
-    })
-
-    test('should render custom label', async () => {
-      await page.goto(url.create)
-      const label = page.locator('label.custom-label[for="field-customLabel"]')
-      await expect(label).toHaveText('#label')
-    })
-
-    test('should render custom error', async () => {
-      await page.goto(url.create)
-      const input = page.locator('input[id="field-customError"]')
-      await input.fill('ab')
-      await expect(input).toHaveValue('ab')
-      const error = page.locator('.custom-error:near(input[id="field-customError"])')
-      const submit = page.locator('button[type="button"][id="action-save"]')
-      await submit.click()
-      await expect(error).toHaveText('#custom-error')
-    })
-
-    test('should render beforeInput and afterInput', async () => {
-      await page.goto(url.create)
-      const input = page.locator('input[id="field-beforeAndAfterInput"]')
-
-      const prevSibling = await input.evaluateHandle((el) => {
-        return el.previousElementSibling
-      })
-      const prevSiblingText = await page.evaluate((el) => el.textContent, prevSibling)
-      expect(prevSiblingText).toEqual('#before-input')
-
-      const nextSibling = await input.evaluateHandle((el) => {
-        return el.nextElementSibling
-      })
-      const nextSiblingText = await page.evaluate((el) => el.textContent, nextSibling)
-      expect(nextSiblingText).toEqual('#after-input')
-    })
-
-    test('should create hasMany with multiple texts', async () => {
-      const input = 'five'
-      const furtherInput = 'six'
-
-      await page.goto(url.create)
-      const requiredField = page.locator('#field-text')
-      const field = page.locator('.field-hasMany')
-
-      await requiredField.fill(String(input))
-      await field.click()
-      await page.keyboard.type(input)
-      await page.keyboard.press('Enter')
-      await page.keyboard.type(furtherInput)
-      await page.keyboard.press('Enter')
-      await saveDocAndAssert(page)
-      await expect(field.locator('.rs__value-container')).toContainText(input)
-      await expect(field.locator('.rs__value-container')).toContainText(furtherInput)
-    })
-  })
-
-  describe('number', () => {
-    let url: AdminUrlUtil
-    beforeAll(() => {
-      url = new AdminUrlUtil(serverURL, 'number-fields')
-    })
-
-    test('should display field in list view', async () => {
-      await page.goto(url.list)
-      const textCell = page.locator('.row-1 .cell-number')
-      await expect(textCell).toHaveText(String(numberDoc.number))
-    })
-
-    test('should filter Number fields in the collection view - greaterThanOrEqual', async () => {
-      await page.goto(url.list)
-
-      // should have 3 entries
-      await expect(page.locator('table >> tbody >> tr')).toHaveCount(3)
-
-      // open the filter options
-      await page.locator('.list-controls__toggle-where').click()
-      await expect(page.locator('.list-controls__where.rah-static--height-auto')).toBeVisible()
-      await page.locator('.where-builder__add-first-filter').click()
-
-      const initialField = page.locator('.condition__field')
-      const operatorField = page.locator('.condition__operator')
-      const valueField = page.locator('.condition__value >> input')
-
-      // select Number field to filter on
-      await initialField.click()
-      const initialFieldOptions = initialField.locator('.rs__option')
-      await initialFieldOptions.locator('text=number').first().click()
-      await expect(initialField.locator('.rs__single-value')).toContainText('Number')
-
-      // select >= operator
-      await operatorField.click()
-      const operatorOptions = operatorField.locator('.rs__option')
-      await operatorOptions.last().click()
-      await expect(operatorField.locator('.rs__single-value')).toContainText(
-        'is greater than or equal to',
-      )
-
-      // enter value of 3
-      await valueField.fill('3')
-      await expect(valueField).toHaveValue('3')
-      await wait(300)
-
-      // should have 2 entries after filtering
-      await expect(page.locator('table >> tbody >> tr')).toHaveCount(2)
-    })
-
-    test('should create', async () => {
-      const input = 5
-
-      await page.goto(url.create)
-      const field = page.locator('#field-number')
-      await field.fill(String(input))
-      await saveDocAndAssert(page)
-      await expect(field).toHaveValue(String(input))
-    })
-
-    test('should create hasMany', async () => {
-      const input = 5
-
-      await page.goto(url.create)
-      const field = page.locator('.field-hasMany')
-      await field.click()
-      await page.keyboard.type(String(input))
-      await page.keyboard.press('Enter')
-      await saveDocAndAssert(page)
-      await expect(field.locator('.rs__value-container')).toContainText(String(input))
-    })
-
-    test('should bypass min rows validation when no rows present and field is not required', async () => {
-      await page.goto(url.create)
-      await saveDocAndAssert(page)
-      await expect(page.locator('.Toastify')).toContainText('successfully')
-    })
-
-    test('should fail min rows validation when rows are present', async () => {
-      const input = 5
-
-      await page.goto(url.create)
-      await page.locator('.field-withMinRows').click()
-
-      await page.keyboard.type(String(input))
-      await page.keyboard.press('Enter')
-      await page.click('#action-save', { delay: 100 })
-
-      await expect(page.locator('.Toastify')).toContainText(
-        'The following field is invalid: withMinRows',
-      )
-    })
   })
 
   describe('indexed', () => {
@@ -449,107 +200,6 @@ describe('fields', () => {
     })
   })
 
-  describe('point', () => {
-    let url: AdminUrlUtil
-    let filledGroupPoint
-    let emptyGroupPoint
-    beforeEach(async () => {
-      url = new AdminUrlUtil(serverURL, pointFieldsSlug)
-      filledGroupPoint = await payload.create({
-        collection: pointFieldsSlug,
-        data: {
-          group: { point: [4, 2] },
-          localized: [4, 2],
-          point: [5, 5],
-        },
-      })
-      emptyGroupPoint = await payload.create({
-        collection: pointFieldsSlug,
-        data: {
-          group: {},
-          localized: [3, -2],
-          point: [5, 5],
-        },
-      })
-    })
-
-    test('should save point', async () => {
-      await page.goto(url.create)
-      const longField = page.locator('#field-longitude-point')
-      await longField.fill('9')
-
-      const latField = page.locator('#field-latitude-point')
-      await latField.fill('-2')
-
-      const localizedLongField = page.locator('#field-longitude-localized')
-      await localizedLongField.fill('1')
-
-      const localizedLatField = page.locator('#field-latitude-localized')
-      await localizedLatField.fill('-1')
-
-      const groupLongitude = page.locator('#field-longitude-group__point')
-      await groupLongitude.fill('3')
-
-      const groupLatField = page.locator('#field-latitude-group__point')
-      await groupLatField.fill('-8')
-
-      await saveDocAndAssert(page)
-      await expect(longField).toHaveAttribute('value', '9')
-      await expect(latField).toHaveAttribute('value', '-2')
-      await expect(localizedLongField).toHaveAttribute('value', '1')
-      await expect(localizedLatField).toHaveAttribute('value', '-1')
-      await expect(groupLongitude).toHaveAttribute('value', '3')
-      await expect(groupLatField).toHaveAttribute('value', '-8')
-    })
-
-    test('should update point', async () => {
-      await page.goto(url.edit(emptyGroupPoint.id))
-      await page.waitForURL(`**/${emptyGroupPoint.id}`)
-      const longField = page.locator('#field-longitude-point')
-      await longField.fill('9')
-
-      const latField = page.locator('#field-latitude-point')
-      await latField.fill('-2')
-
-      const localizedLongField = page.locator('#field-longitude-localized')
-      await localizedLongField.fill('2')
-
-      const localizedLatField = page.locator('#field-latitude-localized')
-      await localizedLatField.fill('-2')
-
-      const groupLongitude = page.locator('#field-longitude-group__point')
-      await groupLongitude.fill('3')
-
-      const groupLatField = page.locator('#field-latitude-group__point')
-      await groupLatField.fill('-8')
-
-      await saveDocAndAssert(page)
-
-      await expect(longField).toHaveAttribute('value', '9')
-      await expect(latField).toHaveAttribute('value', '-2')
-      await expect(localizedLongField).toHaveAttribute('value', '2')
-      await expect(localizedLatField).toHaveAttribute('value', '-2')
-      await expect(groupLongitude).toHaveAttribute('value', '3')
-      await expect(groupLatField).toHaveAttribute('value', '-8')
-    })
-
-    test('should be able to clear a value point', async () => {
-      await page.goto(url.edit(filledGroupPoint.id))
-      await page.waitForURL(`**/${filledGroupPoint.id}`)
-
-      const groupLongitude = page.locator('#field-longitude-group__point')
-      await groupLongitude.fill('')
-
-      const groupLatField = page.locator('#field-latitude-group__point')
-      await groupLatField.fill('')
-
-      await saveDocAndAssert(page)
-
-      await expect(groupLongitude).toHaveAttribute('value', '')
-      await expect(groupLatField).toHaveAttribute('value', '')
-    })
-  })
-
   describe('collapsible', () => {
     let url: AdminUrlUtil
     beforeAll(() => {
@@ -633,317 +283,6 @@ describe('fields', () => {
       await page.goto(url.create)
       const field = page.locator('#field-disableSort .collapsible__drag')
       expect(await field.count()).toEqual(0)
-    })
-  })
-
-  describe('tabs', () => {
-    let url: AdminUrlUtil
-    beforeAll(() => {
-      url = new AdminUrlUtil(serverURL, tabsFieldsSlug)
-    })
-
-    test('should fill and retain a new value within a tab while switching tabs', async () => {
-      const textInRowValue = 'hello'
-      const numberInRowValue = '23'
-      const jsonValue = '{ "foo": "bar"}'
-
-      await page.goto(url.create)
-      await page.waitForURL(url.create)
-
-      await switchTab(page, '.tabs-field__tab-button:has-text("Tab with Row")')
-      await page.locator('#field-textInRow').fill(textInRowValue)
-      await page.locator('#field-numberInRow').fill(numberInRowValue)
-      await page.locator('.json-field .inputarea').fill(jsonValue)
-
-      await wait(300)
-
-      await switchTab(page, '.tabs-field__tab-button:has-text("Tab with Array")')
-      await switchTab(page, '.tabs-field__tab-button:has-text("Tab with Row")')
-
-      await expect(page.locator('#field-textInRow')).toHaveValue(textInRowValue)
-      await expect(page.locator('#field-numberInRow')).toHaveValue(numberInRowValue)
-      await expect(page.locator('.json-field .lines-content')).toContainText(jsonValue)
-    })
-
-    test('should retain updated values within tabs while switching between tabs', async () => {
-      const textInRowValue = 'new value'
-      const jsonValue = '{ "new": "value"}'
-      await page.goto(url.list)
-      await navigateToListCellLink(page)
-
-      // Go to Row tab, update the value
-      await switchTab(page, '.tabs-field__tab-button:has-text("Tab with Row")')
-
-      await page.locator('#field-textInRow').fill(textInRowValue)
-      await page.locator('.json-field .inputarea').fill(jsonValue)
-
-      await wait(500)
-
-      // Go to Array tab, then back to Row. Make sure new value is still there
-      await switchTab(page, '.tabs-field__tab-button:has-text("Tab with Array")')
-      await switchTab(page, '.tabs-field__tab-button:has-text("Tab with Row")')
-
-      await expect(page.locator('#field-textInRow')).toHaveValue(textInRowValue)
-      await expect(page.locator('.json-field .lines-content')).toContainText(jsonValue)
-
-      // Go to array tab, save the doc
-      await switchTab(page, '.tabs-field__tab-button:has-text("Tab with Array")')
-      await saveDocAndAssert(page)
-
-      // Go back to row tab, make sure the new value is still present
-      await switchTab(page, '.tabs-field__tab-button:has-text("Tab with Row")')
-      await expect(page.locator('#field-textInRow')).toHaveValue(textInRowValue)
-    })
-
-    test('should render array data within unnamed tabs', async () => {
-      await page.goto(url.list)
-      await navigateToListCellLink(page)
-      await switchTab(page, '.tabs-field__tab-button:has-text("Tab with Array")')
-      await expect(page.locator('#field-array__0__text')).toHaveValue("Hello, I'm the first row")
-    })
-
-    test('should render array data within named tabs', async () => {
-      await page.goto(url.list)
-      await navigateToListCellLink(page)
-      await switchTab(page, '.tabs-field__tab-button:nth-child(5)')
-      await expect(page.locator('#field-tab__array__0__text')).toHaveValue(
-        "Hello, I'm the first row, in a named tab",
-      )
-    })
-  })
-
-  describe('date', () => {
-    let url: AdminUrlUtil
-    beforeAll(() => {
-      url = new AdminUrlUtil(serverURL, 'date-fields')
-    })
-
-    test('should display formatted date in list view table cell', async () => {
-      await page.goto(url.list)
-      const formattedDateCell = page.locator('.row-1 .cell-timeOnly')
-      await expect(formattedDateCell).toContainText(' Aug ')
-
-      const notFormattedDateCell = page.locator('.row-1 .cell-default')
-      await expect(notFormattedDateCell).toContainText('August')
-    })
-
-    test('should display formatted date in useAsTitle', async () => {
-      await page.goto(url.list)
-      await page.locator('.row-1 .cell-default a').click()
-      await expect(page.locator('.doc-header__title.render-title')).toContainText('August')
-    })
-
-    test('should clear date', async () => {
-      await page.goto(url.create)
-      const dateField = page.locator('#field-default input')
-      await expect(dateField).toBeVisible()
-      await dateField.fill('02/07/2023')
-      await expect(dateField).toHaveValue('02/07/2023')
-      await saveDocAndAssert(page)
-
-      const clearButton = page.locator('#field-default .date-time-picker__clear-button')
-      await expect(clearButton).toBeVisible()
-      await clearButton.click()
-      await expect(dateField).toHaveValue('')
-    })
-
-    describe('localized dates', () => {
-      describe('EST', () => {
-        test.use({
-          geolocation: {
-            latitude: 42.3314,
-            longitude: -83.0458,
-          },
-          timezoneId: 'America/Detroit',
-        })
-        test('create EST day only date', async () => {
-          await page.goto(url.create)
-          await page.waitForURL(`**/${url.create}`)
-          const dateField = page.locator('#field-default input')
-
-          // enter date in default date field
-          await dateField.fill('02/07/2023')
-          await saveDocAndAssert(page)
-
-          // get the ID of the doc
-          const routeSegments = page.url().split('/')
-          const id = routeSegments.pop()
-
-          // fetch the doc (need the date string from the DB)
-          const { doc } = await client.findByID({ id, auth: true, slug: 'date-fields' })
-
-          expect(doc.default).toEqual('2023-02-07T12:00:00.000Z')
-        })
-      })
-
-      describe('PST', () => {
-        test.use({
-          geolocation: {
-            latitude: 37.774929,
-            longitude: -122.419416,
-          },
-          timezoneId: 'America/Los_Angeles',
-        })
-
-        test('create PDT day only date', async () => {
-          await page.goto(url.create)
-          await page.waitForURL(`**/${url.create}`)
-          const dateField = page.locator('#field-default input')
-
-          // enter date in default date field
-          await dateField.fill('02/07/2023')
-          await saveDocAndAssert(page)
-
-          // get the ID of the doc
-          const routeSegments = page.url().split('/')
-          const id = routeSegments.pop()
-
-          // fetch the doc (need the date string from the DB)
-          const { doc } = await client.findByID({ id, auth: true, slug: 'date-fields' })
-
-          expect(doc.default).toEqual('2023-02-07T12:00:00.000Z')
-        })
-      })
-
-      describe('ST', () => {
-        test.use({
-          geolocation: {
-            latitude: -14.5994,
-            longitude: -171.857,
-          },
-          timezoneId: 'Pacific/Apia',
-        })
-
-        test('create ST day only date', async () => {
-          await page.goto(url.create)
-          await page.waitForURL(`**/${url.create}`)
-          const dateField = page.locator('#field-default input')
-
-          // enter date in default date field
-          await dateField.fill('02/07/2023')
-          await saveDocAndAssert(page)
-
-          // get the ID of the doc
-          const routeSegments = page.url().split('/')
-          const id = routeSegments.pop()
-
-          // fetch the doc (need the date string from the DB)
-          const { doc } = await client.findByID({ id, auth: true, slug: 'date-fields' })
-
-          expect(doc.default).toEqual('2023-02-07T12:00:00.000Z')
-        })
-      })
-    })
-  })
-
-  describe('upload', () => {
-    let url: AdminUrlUtil
-    beforeAll(() => {
-      url = new AdminUrlUtil(serverURL, 'uploads')
-    })
-
-    async function uploadImage() {
-      await page.goto(url.create)
-
-      // create a jpg upload
-      await page
-        .locator('.file-field__upload input[type="file"]')
-        .setInputFiles(path.resolve(dirname, './collections/Upload/payload.jpg'))
-      await expect(page.locator('.file-field .file-field__filename')).toHaveValue('payload.jpg')
-      await saveDocAndAssert(page)
-    }
-
-    // eslint-disable-next-line playwright/expect-expect
-    test('should upload files', async () => {
-      await uploadImage()
-    })
-
-    // test that the image renders
-    test('should render uploaded image', async () => {
-      await uploadImage()
-      await expect(page.locator('.file-field .file-details img')).toHaveAttribute(
-        'src',
-        '/api/uploads/file/payload-1.jpg',
-      )
-    })
-
-    test('should upload using the document drawer', async () => {
-      await uploadImage()
-      await wait(1000)
-      // Open the media drawer and create a png upload
-
-      await openDocDrawer(page, '.field-type.upload .upload__toggler.doc-drawer__toggler')
-
-      await page
-        .locator('[id^=doc-drawer_uploads_1_] .file-field__upload input[type="file"]')
-        .setInputFiles(path.resolve(dirname, './uploads/payload.png'))
-      await expect(
-        page.locator('[id^=doc-drawer_uploads_1_] .file-field__upload .file-field__filename'),
-      ).toHaveValue('payload.png')
-      await page.locator('[id^=doc-drawer_uploads_1_] #action-save').click()
-      await expect(page.locator('.Toastify')).toContainText('successfully')
-
-      // Assert that the media field has the png upload
-      await expect(
-        page.locator('.field-type.upload .file-details .file-meta__url a'),
-      ).toHaveAttribute('href', '/api/uploads/file/payload-1.png')
-      await expect(
-        page.locator('.field-type.upload .file-details .file-meta__url a'),
-      ).toContainText('payload-1.png')
-      await expect(page.locator('.field-type.upload .file-details img')).toHaveAttribute(
-        'src',
-        '/api/uploads/file/payload-1.png',
-      )
-      await saveDocAndAssert(page)
-    })
-
-    test('should clear selected upload', async () => {
-      await uploadImage()
-      await wait(1000) // TODO: Fix this. Need to wait a bit until the form in the drawer mounted, otherwise values sometimes disappear. This is an issue for all drawers
-
-      await openDocDrawer(page, '.field-type.upload .upload__toggler.doc-drawer__toggler')
-
-      await wait(1000)
-
-      await page
-        .locator('[id^=doc-drawer_uploads_1_] .file-field__upload input[type="file"]')
-        .setInputFiles(path.resolve(dirname, './uploads/payload.png'))
-      await expect(
-        page.locator('[id^=doc-drawer_uploads_1_] .file-field__upload .file-field__filename'),
-      ).toHaveValue('payload.png')
-      await page.locator('[id^=doc-drawer_uploads_1_] #action-save').click()
-      await expect(page.locator('.Toastify')).toContainText('successfully')
-      await page.locator('.field-type.upload .file-details__remove').click()
-    })
-
-    test('should select using the list drawer and restrict mimetype based on filterOptions', async () => {
-      await uploadImage()
-
-      await openDocDrawer(page, '.field-type.upload .upload__toggler.list-drawer__toggler')
-
-      const jpgImages = page.locator('[id^=list-drawer_1_] .upload-gallery img[src$=".jpg"]')
-      await expect
-        .poll(async () => await jpgImages.count(), { timeout: POLL_TOPASS_TIMEOUT })
-        .toEqual(0)
-    })
-
-    test.skip('should show drawer for input field when enableRichText is false', async () => {
-      const uploads3URL = new AdminUrlUtil(serverURL, 'uploads3')
-      await page.goto(uploads3URL.create)
-
-      // create file in uploads 3 collection
-      await page
-        .locator('.file-field__upload input[type="file"]')
-        .setInputFiles(path.resolve(dirname, './collections/Upload/payload.jpg'))
-      await expect(page.locator('.file-field .file-field__filename')).toContainText('payload.jpg')
-      await page.locator('#action-save').click()
-
-      await wait(200)
-
-      // open drawer
-      await openDocDrawer(page, '.field-type.upload .list-drawer__toggler')
-      // check title
-      await expect(page.locator('.list-drawer__header-text')).toContainText('Uploads 3')
     })
   })
